Feature/migrate to GitHub actions (#109)
* Added an AGENTS.md file in order to allow AI Agents to assist with upkeep.
* Updated to migrate to github actions in order to build.
* Updated the StringExtended to ignore in phpcs

---------

+++ src/Mhorninger/MySQLite/MySQL/Miscellaneous.php
@@ -9,7 +9,7 @@ trait Miscellaneous
     {
         //phpcs:enable
         if ($numeric && is_numeric($numeric)) {
-            return long2ip(sprintf('%d', $numeric));
+            return long2ip((int) sprintf('%d', $numeric));
         }
     }
 }
